move test-support files into /spec/suppoport
